refactor: separate file to more readable and review easier
